added 2

+++ lab2/cmd/file.go
@@ -1,7 +1,7 @@
 package main
 
-func merge(leftPart []float64, rightPart []float64) []float64 {
-	var mergedArray []float64
+func merge(leftPart []int, rightPart []int) []int {
+	var mergedArray []int
 	l := 0
 	r := 0
 	for l < len(leftPart) && r < len(rightPart) {
@@ -19,7 +19,7 @@ func merge(leftPart []float64, rightPart []float64) []float64 {
 	return mergedArray
 }
 
-func mergeSort(array []float64) []float64 {
+func mergeSort(array []int) []int {
 	if len(array) < 2 {
 		return array
 	}
@@ -29,7 +29,7 @@ func mergeSort(array []float64) []float64 {
 	return merge(leftPart, rightPart)
 }
 
-func findKthLargest(array []float64, neededNum int) []float64 {
+func findKthLargest(array []int, neededNum int) []int {
 	NewArray := mergeSort(array)
 	return NewArray
 }
